% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{osmFISH_phenotype_simu}
\alias{osmFISH_phenotype_simu}
\title{Simulated Phenotype Labels for osmFISH Dataset}
\format{
A character vector of length 2. Each element is the name of a region selected from
the `osmFISH_metadata_region` object. These two regions are used to simulate binary phenotypic classes.
}
\source{
Extracted from simulated osmFISH region labels in the SpaPheno tutorial.
}
\usage{
data(osmFISH_phenotype_simu)
}
\description{
A character vector containing the names of two spatial regions used as simulated binary phenotypes
in the osmFISH synthetic dataset. These phenotype labels define the two groups (e.g., "Tumor" vs "Normal")
for training and evaluating spatially-informed predictive models.
}
\examples{
data(osmFISH_phenotype_simu)
print(osmFISH_phenotype_simu)

}
\seealso{
\code{\link{osmFISH_metadata_region}}, \code{\link{generate_simulated_bulk_data}}
}
\keyword{datasets}
