% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaPheno_SHAP_waterfall_plot.R
\name{SpaPheno_SHAP_waterfall_plot}
\alias{SpaPheno_SHAP_waterfall_plot}
\title{SHAP Waterfall Plot for a Single Sample}
\usage{
SpaPheno_SHAP_waterfall_plot(shap_df, sample_id, top_n = 10)
}
\arguments{
\item{shap_df}{A data.frame of SHAP results. Must include:
\code{sample}, \code{feature}, and \code{phi} columns.}

\item{sample_id}{A string. The ID of the sample to visualize (must match \code{sample} in \code{shap_df}).}

\item{top_n}{Integer; number of top features to include. Default is 10.}
}
\value{
A \code{ggplot2} object displaying the SHAP waterfall plot.
}
\description{
Generate a SHAP waterfall plot to visualize the most influential features
contributing to a model's prediction for a given sample. Features are ordered
by the magnitude of their SHAP values, and color-coded by direction (positive/negative).
}
\details{
- Filters SHAP values for the selected \code{sample_id}.
- Selects top \code{top_n} features based on absolute SHAP value.
- Plots bars horizontally using \code{ggplot2}, colored by contribution sign.
}
\examples{
\dontrun{

data("osmFISH_metadata_cellType")
data("osmFISH_bulk_decon")
data("osmFISH_bulk_pheno")
data("osmFISH_bulk_coordinate")

PhenoResult <- SpatialPhenoMap(
  bulk_decon = osmFISH_bulk_decon,
  bulk_pheno = osmFISH_bulk_pheno,
  family = "binomial",
  coord = osmFISH_bulk_coordinate,
  resolution = "single_cell",
  sample_information_cellType = osmFISH_metadata_cellType
)

pred_result <- PhenoResult$pred_score
phenoPlus <- row.names(pred_result[pred_result$label \%in\% "phenotype+", ])
model <- PhenoResult$model
X <- as.data.frame(PhenoResult$cell_type_distribution[phenoPlus, ])

# This step took a very long time
shap_test_plus <- compute_shap_spatial(
  model = model,
  X_bulk = as.data.frame(osmFISH_bulk_decon),
  y_bulk = osmFISH_bulk_pheno,
  X_spatial = X)

SpaPheno_SHAP_waterfall_plot(shap_test_plus, sample_id = "cell_3477", top_n = 10)

}

}
\author{
Bin Duan
}
